Google webfonts integration to customizer.
Functions mostly. Sometimes the firing script to get a new preview
doesn't fire. Perhaps it could be manually called? Added a variety of
functions. Created a custom-control for the webfont selectboxes only to
add a data-details element to the span. If I could find how to add that
to existing format, I think it would be much clearner that way. Please
insepct and do some code review. Still need to decide if how things
have been separated out is the best method. Also want to talk about
having a globabl typography box, and isolated typographies for
headers/branding to give more granularity to these controls.